Fixed autosave

+++ app/src/main/jni/qsp-5.7/src/bindings/android/android.h
@@ -22,7 +22,7 @@
 #define QSP_ANDROIDDEFINES
 
 #ifdef _UNICODE
-typedef int QSP_CHAR;
+typedef unsigned short QSP_CHAR;
 #endif
 
 QSP_EXTERN char *qspW2C(QSP_CHAR *);
